+++ src/parameters.h
@@ -30,7 +30,7 @@ typedef struct {
 
     double * time_step; int n_time_step;
     int enforce_broadband; 
-    double enforce_broadband_kmax;
+    int enforce_broadband_kmax;
     int UseFFTW;
     int NprocY;
 } Parameters;
